Clean up billing stores, remove orphaned code
