+++ src/components/admin/Billing/PricingTierDetails.tsx
@@ -5,7 +5,7 @@ import { Skeleton, Typography } from '@mui/material';
 import { FormattedMessage } from 'react-intl';
 
 import { useTenantUsesExternalPayment } from 'src/context/fetcher/TenantBillingDetails';
-import { useBillingStore } from 'src/stores/Billing/Store';
+import { useBillingStore } from 'src/stores/Billing';
 import { useTenantStore } from 'src/stores/Tenant/Store';
 
 function PricingTierDetails() {
