+++ src/components/tables/Billing/index.tsx
@@ -10,8 +10,10 @@ import Rows from 'src/components/tables/Billing/Rows';
 import EntityTableBody from 'src/components/tables/EntityTable/TableBody';
 import EntityTableHeader from 'src/components/tables/EntityTable/TableHeader';
 import { getTableHeaderWithoutHeaderColor } from 'src/context/Theme';
-import { useBilling_selectedInvoice } from 'src/stores/Billing/hooks';
-import { useBillingStore } from 'src/stores/Billing/Store';
+import {
+    useBilling_selectedInvoice,
+    useBillingStore,
+} from 'src/stores/Billing';
 import { TableStatuses } from 'src/types';
 import { invoiceId } from 'src/utils/billing-utils';
