+++ src/context/Zustand/invariableStores.ts
@@ -12,7 +12,6 @@ import {
 } from 'src/stores/names';
 import { createSchemaEvolutionStore } from 'src/stores/SchemaEvolution/Store';
 import { createShardDetailStore } from 'src/stores/ShardDetail/Store';
-import { createBillingTableStore } from 'src/stores/Tables/Billing/Store';
 import { createSelectableTableStore } from 'src/stores/Tables/Store';
 import { createTransformationCreateStore } from 'src/stores/TransformationCreate/Store';
 import { MessagePrefixes } from 'src/types';
@@ -63,9 +62,6 @@ const invariableStores = {
     ),
     [SelectTableStoreNames.ACCESS_GRANTS_USERS]: createSelectableTableStore(
         SelectTableStoreNames.ACCESS_GRANTS_USERS
-    ),
-    [SelectTableStoreNames.BILLING]: createBillingTableStore(
-        SelectTableStoreNames.BILLING
     ),
     [SelectTableStoreNames.CAPTURE]: createSelectableTableStore(
         SelectTableStoreNames.CAPTURE
