--- src/stores/Billing/hooks.ts
@@ -1,13 +1,0 @@
-import { useBillingStore } from 'src/stores/Billing/Store';
-import { invoiceId } from 'src/utils/billing-utils';
-
-// Selector Hooks
-export const useBilling_selectedInvoice = () => {
-    return useBillingStore((state) =>
-        state.selectedInvoiceId
-            ? (state.invoices.find(
-                  (inv) => invoiceId(inv) === state.selectedInvoiceId
-              ) ?? null)
-            : null
-    );
-};
